Initial commit.
Adds a simple proof-of-concept console app doing a couple simulation
rumble effects. Basically a engine start, stop, revving up to what
could be a "cruise speed rpm", and also revving up to maximum,
with options to come back to idle speeds.

This can probably be better if the normal rpm operation actually
resulted in a much lower feedback rumble, with more feeling in the
extremes (low rpm and very high).