Creates the ForceRumbleback SCS Plugin project.
Also adds SCS SDK headers to the solution.
Introduces VS's editorconfig to try to avoid mixing up tabs and spaces
in source code (placed at solution root).

The SCS Plugin is still not functional, it basically prints out state
changes but does not interface with the controller in any way.

But at least it registers successfully in the game (tested with ATS).