Increases polling delay in case commands fail.
Adds status returns from sendforce() function and checks it to tell
whether sending commands fails and increases the polling delay to avoid
flooding the driver/bus in case the controller is not usable.

In failing polls, it will always retry to send the force-feedback
command at least once to restore effects.

Changed the sendforce() and sendforceWait() functions to reflect global
values as, respectively, SendForce() and SendForceWait().

The poll delay when a force send command fails is increased by 1 second
to whatever current polling delay is so, currently, to 1010ms.