Reviews rpm2force conversion and adds drag accounting.
Reviews the current truck RPM to force conversion for more realistic
feel. Especially in torque plateau interval of RPM, the effect is much
weaker now to simulate a well balanced engine vibration.

Adds a drag computation system to add vibration intensity to reflect how
the engine is struggling to attain a given speed (in rpm). That is, if
the engine is loaded (truck is heavy, uphill) and the gas pedal
depressed hard, it will vibrate more than if cruising in a flat road.

The result would be a great candidate for a realistic turbo gauge.