Adjusts logging not to require a file, and log to game console.
New functions were made to wrap the old log file writing in preprocessor
conditions, and all logging in the plugin was reviewed so that it now
logs to the game console and game.log.txt.

In order to re-enable /var/log/ForceRumbleback.log, a recompilation of
the project is required after enabling the preprocessor flag.