Poller: applies new force of 0 regardless of granularity.
If at some point the plugin tried to stop the vibration effect and it
was weak enough, the update wouldn't be sent to the joystick, causing it
to still vibrate, for instance, while in menu or when turning off the
truck.

With this, if a force of 0 is requested, it will only not update if the
current force is already exactly at 0, still avoiding unnecessary updates.